move css-data to a package
