+++ apps/designer/app/designer/features/style-panel/shared/get-inherited-style.ts
@@ -1,9 +1,9 @@
-import {
-  properties,
-  type StyleValue,
-  type StyleProperty,
+import { properties } from "@webstudio-is/css-data";
+import type {
+  StyleValue,
+  StyleProperty,
+  Instance,
 } from "@webstudio-is/react-sdk";
-import { type Instance } from "@webstudio-is/react-sdk";
 
 export type InheritedStyle = {
   [property in StyleProperty]: {
