+++ packages/css-data/README.md
@@ -1,0 +1,3 @@
+# CSS Data
+
+Generated configs and collections from mdn-data into consumable form.
