+++ packages/css-data/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "@webstudio-is/css-data",
+  "version": "0.1.0",
+  "description": "CSS Data",
+  "author": "Webstudio <github@webstudio.is>",
+  "homepage": "https://webstudio.is",
+  "scripts": {
+    "typecheck": "tsc --noEmit",
+    "checks": "yarn typecheck && yarn lint",
+    "dev": "tsup --watch",
+    "build": "rm -fr lib tsconfig.tsbuildinfo && tsc",
+    "build:mdn-data": "tsx ./bin/mdn-data.ts ./src",
+    "lint": "eslint ./src --ext .ts,.tsx --max-warnings 0"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "@webstudio-is/jest-config": "*",
+    "mdn-data": "2.0.23",
+    "tsup": "^6.1.3",
+    "typescript": "4.7.4",
+    "@types/css-tree": "^1.0.7",
+    "css-tree": "^2.1.0",
+    "camelcase": "^6.3.0"
+  },
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "files": [
+    "lib/*",
+    "README.md",
+    "!*.test.*"
+  ],
+  "license": "MIT",
+  "private": false,
+  "sideEffects": false,
+  "tsup": {
+    "entry": [
+      "src/index.ts"
+    ],
+    "format": "cjs",
+    "outDir": "lib"
+  }
+}
