+++ packages/react-sdk/package.json
@@ -7,7 +7,6 @@
   "scripts": {
     "dev": "tsup --watch",
     "build": "yarn build:lib",
-    "build:mdn-data": "tsx ./bin/mdn-data.ts ./src/css",
     "build:args": "tsx ./bin/generate-arg-types.ts './src/components/*.tsx !./src/**/*.stories.tsx !./src/**/*.ws.tsx'",
     "typecheck": "tsc --noEmit",
     "build:lib": "rm -fr lib tsconfig.tsbuildinfo && tsc",
@@ -32,17 +31,13 @@
     "@storybook/manager-webpack4": "^6.5.6",
     "@storybook/react": "^6.5.6",
     "@storybook/testing-library": "^0.0.11",
-    "@types/css-tree": "^1.0.7",
     "@types/fs-extra": "^9.0.13",
     "@types/node": "^17.0.21",
     "@webstudio-is/jest-config": "*",
     "@webstudio-is/prisma-client": "*",
     "babel-loader": "^8.2.5",
-    "camelcase": "^6.3.0",
-    "css-tree": "^2.1.0",
     "esbuild": "^0.14.25",
     "esbuild-node-externals": "^1.4.1",
-    "mdn-data": "2.0.23",
     "react": "^17.0.2",
     "react-docgen-typescript": "^2.2.2",
     "react-dom": "^17.0.2",
@@ -62,6 +57,7 @@
     "@webstudio-is/asset-uploader": "^*",
     "@webstudio-is/fonts": "*",
     "@webstudio-is/icons": "*",
+    "@webstudio-is/css-data": "*",
     "@webstudio-is/jest-config": "*",
     "immer": "^9.0.12",
     "mitt": "^3.0.0",
