+++ packages/react-sdk/tsconfig.json
@@ -25,6 +25,10 @@
     },
 
     {
+      "path": "../css-data"
+    },
+
+    {
       "path": "../fonts"
     },
